requirements.txt
discord.py~=2.3
aiosqlite~=0.19
aiohttp~=3.9
python-dotenv~=1.0
